Replace C-style casts with C++-style casts everywhere

+++ src/house.h
@@ -223,7 +223,7 @@ class House
 			return bedsList;
 		}
 		uint32_t getBedCount() {
-			return static_cast<uint32_t>(std::ceil((double)bedsList.size() / 2));   //each bed takes 2 sqms of space, ceil is just for bad maps
+			return static_cast<uint32_t>(std::ceil(bedsList.size() / 2.));   //each bed takes 2 sqms of space, ceil is just for bad maps
 		}
 
 	private:
